Move log helper functions to host.go
